refactor(output2graphs): head と base のグラフを出力する処理を関数化
